+++ addOns/automation/src/main/java/org/zaproxy/addon/automation/jobs/PolicyDefinition.java
@@ -22,6 +22,7 @@ package org.zaproxy.addon.automation.jobs;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 import lombok.Getter;
 import lombok.Setter;
 import org.parosproxy.paros.Constant;
@@ -31,12 +32,14 @@ import org.parosproxy.paros.core.scanner.Plugin.AttackStrength;
 import org.parosproxy.paros.core.scanner.PluginFactory;
 import org.zaproxy.addon.automation.AutomationData;
 import org.zaproxy.addon.automation.AutomationProgress;
-import org.zaproxy.addon.automation.jobs.PolicyDefinition.Rule;
 import org.zaproxy.zap.extension.ascan.ScanPolicy;
 
 @Getter
 @Setter
 public class PolicyDefinition extends AutomationData {
+
+    private static final String DEFAULT_STRENGTH_KEY = "defaultStrength";
+    private static final String DEFAULT_THRESHOLD_KEY = "defaultThreshold";
 
     protected static final String RULES_ELEMENT_NAME = "rules";
 
@@ -44,23 +47,34 @@ public class PolicyDefinition extends AutomationData {
     private String defaultThreshold = JobUtils.thresholdToI18n(AlertThreshold.MEDIUM.name());
     private List<Rule> rules = new ArrayList<>();
 
-    public static void parsePolicyDefinition(
-            Object policyDefnObj,
-            PolicyDefinition policyDefinition,
-            String jobName,
-            AutomationProgress progress) {
+    public void parsePolicyDefinition(
+            Object policyDefnObj, String jobName, AutomationProgress progress) {
 
+        if (policyDefnObj == null) {
+            this.defaultStrength = null;
+            return;
+        }
         if (policyDefnObj instanceof LinkedHashMap<?, ?>) {
-            LinkedHashMap<?, ?> policyDefnData = (LinkedHashMap<?, ?>) policyDefnObj;
+            @SuppressWarnings("unchecked")
+            LinkedHashMap<Object, Object> policyDefnData =
+                    (LinkedHashMap<Object, Object>) policyDefnObj;
+
+            checkAndSetDefault(policyDefnData, DEFAULT_STRENGTH_KEY, AttackStrength.MEDIUM.name());
+            checkAndSetDefault(policyDefnData, DEFAULT_THRESHOLD_KEY, AlertThreshold.MEDIUM.name());
+
+            if (policyDefnData.isEmpty() || undefinedDefinition(policyDefnData)) {
+                this.defaultStrength = null;
+                return;
+            }
 
             JobUtils.applyParamsToObject(
                     policyDefnData,
-                    policyDefinition,
+                    this,
                     jobName,
                     new String[] {PolicyDefinition.RULES_ELEMENT_NAME},
                     progress);
 
-            List<Rule> rules = new ArrayList<>();
+            this.rules = new ArrayList<>();
             ScanPolicy scanPolicy = new ScanPolicy();
             PluginFactory pluginFactory = scanPolicy.getPluginFactory();
 
@@ -89,7 +103,7 @@ public class PolicyDefinition extends AutomationData {
                             if (strength != null) {
                                 rule.setStrength(strength.name().toLowerCase());
                             }
-                            rules.add(rule);
+                            this.rules.add(rule);
 
                         } else {
                             progress.warn(
@@ -98,7 +112,6 @@ public class PolicyDefinition extends AutomationData {
                         }
                     }
                 }
-                policyDefinition.setRules(rules);
             } else if (o != null) {
                 progress.warn(
                         Constant.messages.getString(
@@ -117,7 +130,32 @@ public class PolicyDefinition extends AutomationData {
         }
     }
 
+    private static void checkAndSetDefault(
+            LinkedHashMap<Object, Object> policyDefnData, String key, String value) {
+        if (policyDefnData.containsKey(key) && policyDefnData.get(key) == null) {
+            policyDefnData.put(key, value);
+        }
+    }
+
+    private static boolean undefinedDefinition(Map<?, ?> policyDefnData) {
+        Object rules = policyDefnData.get(RULES_ELEMENT_NAME);
+        boolean rulesInvalid = false;
+        if (rules instanceof List<?>) {
+            rulesInvalid = ((List<?>) rules).isEmpty();
+        } else if ((String) rules == null) {
+            rulesInvalid = true;
+        }
+        return (String) policyDefnData.get(DEFAULT_STRENGTH_KEY) == null
+                && (String) policyDefnData.get(DEFAULT_THRESHOLD_KEY) == null
+                && rulesInvalid;
+    }
+
     public ScanPolicy getScanPolicy(String jobName, AutomationProgress progress) {
+        if (getDefaultStrength() == null) {
+            // Nothing defined
+            return null;
+        }
+
         ScanPolicy scanPolicy = new ScanPolicy();
 
         // Set default strength