+++ addOns/sequence/CHANGELOG.md
@@ -9,8 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `sequence-import` to import HARs as sequences.
   - `sequence-activeScan` to active scan sequences.
 - Data for reporting.
-- Sequence active scan policy.
 - Stats for import automation and active scan.
+- Sequence active scan policy which will be used if neither a policy nor policyDefinition are set.
 
 ### Changed
 - Update minimum ZAP version to 2.15.0.